requirements_test.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-homeassistant-custom-component>=0.13.5
requests-mock>=1.11.0
coverage>=7.0.0
